// src/app/models/NWS_API_Models/gridpointForecast.ts
/**
 * weather.gov API
 * weather.gov API
 *
 * OpenAPI spec version: 1.11.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { GeometryString } from './geometryString';
import { GridpointForecastPeriod } from './gridpointForecastPeriod';
import { GridpointForecastUnits } from './gridpointForecastUnits';
import { ISO8601Interval } from './iSO8601Interval';
import { JsonLdContext } from './jsonLdContext';
import { QuantitativeValue } from './quantitativeValue';

/**
 * A multi-day forecast for a 2.5km grid square.
 */
export interface GridpointForecast {
  context?: JsonLdContext;
  geometry?: GeometryString;
  properties: {
    elevation?: QuantitativeValue;
    forecastGenerator?: string;
    generatedAt?: Date;
    periods?: Array<GridpointForecastPeriod>;

    units?: GridpointForecastUnits;
    updateTime?: Date;
    updated?: Date;
    validTimes?: ISO8601Interval;
  };
}