// src/app/models/current-temp.model.ts
export interface CurrentTemp {
    latitude?: number
    longitude?: number
    generationtime_ms?: number
    utc_offset_seconds?: number
    timezone?: string
    timezone_abbreviation?: string
    elevation?: number
    current_weather?: {
      temperature?: number
      windspeed?: number
      winddirection?: number
      weathercode?: number
      is_day?: number
      time?: string
    }
  }
  